[macOS] vrbo.com: black color extension shows up after searching for homes for the first time https://bugs.webkit.org/show_bug.cgi?id=297735 rdar://158871557
Reviewed by Abrar Rahman Protyasha.

Make another small adjustment to the color sampling heuristic — leverage the ability to blend the
sampled fixed color against the background color of the page (added in 298351@main) to avoid
clamping sampled colors with lighter opacity to opaque colors (with a threshold of 0.75). This
prevents a solid black bar from appearing behind the navigation bar, when the top sampled color is a
color like 40% transparent black (i.e. the background color of the background behind the popup after
searching for homes for the first time on vrbo.com).

* LayoutTests/fast/page-color-sampling/blend-modal-background-color-2-expected.txt: Copied from LayoutTests/fast/page-color-sampling/blend-modal-background-color-expected.txt.
* LayoutTests/fast/page-color-sampling/blend-modal-background-color-2.html: Copied from LayoutTests/fast/page-color-sampling/blend-modal-background-color.html.

Add a second layout test for this color blending heuristic that bypasses the modal dimming view
detection, but whose opacity is low enough that the resulting sampled color should not be clamped to
solid black.

* LayoutTests/fast/page-color-sampling/blend-modal-background-color-expected.txt:
* LayoutTests/fast/page-color-sampling/blend-modal-background-color.html:
* Source/WebCore/page/LocalFrameView.cpp:
(WebCore::LocalFrameView::fixedContainerEdges const):

Canonical link: https://commits.webkit.org/299029@main