[web-animations] implement the new endpoint-inclusive active interval behavior for `Animation.commitStyles()` https://bugs.webkit.org/show_bug.cgi?id=297594 rdar://158684709
Reviewed by Cameron McCormack.

The behavior of `Animation.commitStyles()` changed to use endpoint-inclusive active intervals [0]:

> In order to simplify the common case of persisting a completed animation, the procedure to commit computed styles uses
> endpoint-*inclusive* timing when determining the phase of the animation. As a result, the following code will persist
> the `transform: translateY(100px)` style in specified style despite not having a fill mode of `forwards` or `both`:
>
>
>```
> elem.animate({ transform: 'translateY(100px)' }, 200).finished.then(() => {
>   elem.commitStyles();
> });
>```

To implement this new behavior, we add a new `EndpointInclusiveActiveInterval` boolean enum and a new member of that
type to the `AnimationEffectTiming::ResolutionData` struct used to resolve timing for animation effects. In order to
pass that value in from `Animation::commitStyles()`, we have to pass such a value through `Element::resolve()` →
`KeyframeEffect::apply()` → `AnimationEffect::getComputedTiming()`.

Then within `Animation::commitStyles()` we make a change not to use the `KeyframeEffectStack` associated with
the animation's target as that stack would not contain any effect associated with an animation that is no longer
_relevant_ [1], which would be the case of a finished animation that is not forward-filling. So we now make a sorted
copy of all animations associated with the animation's target and iterate through the associated effects, making
sure to pass `EndpointInclusiveActiveInterval::Yes` when calling `Animation::resolve()` to apply the animation's value.

Note that there are remaining failures in the WPT test which are not relevant to that spec change and which are tracked
by bug 297657.

[0] https://github.com/w3c/csswg-drafts/commit/aa2e02d2daee8d8eb146562f7a84d58f54f4b10c
[1] https://drafts.csswg.org/web-animations-1/#relevant-animations-section

* LayoutTests/imported/w3c/web-platform-tests/web-animations/interfaces/Animation/commitStyles-expected.txt:
* Source/WebCore/animation/AnimationEffect.cpp:
(WebCore::AnimationEffect::resolutionData const):
(WebCore::AnimationEffect::getComputedTiming):
* Source/WebCore/animation/AnimationEffect.h:
* Source/WebCore/animation/AnimationEffectTiming.cpp:
(WebCore::AnimationEffectTiming::getBasicTiming const):
* Source/WebCore/animation/AnimationEffectTiming.h:
* Source/WebCore/animation/CSSTransition.cpp:
(WebCore::CSSTransition::resolve):
* Source/WebCore/animation/CSSTransition.h:
* Source/WebCore/animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::apply):
* Source/WebCore/animation/KeyframeEffect.h:
* Source/WebCore/animation/WebAnimation.cpp:
(WebCore::WebAnimation::resolve):
(WebCore::WebAnimation::commitStyles):
* Source/WebCore/animation/WebAnimation.h:
* Source/WebCore/animation/WebAnimationTypes.h:
* Source/WebCore/platform/animation/AcceleratedEffect.cpp:
(WebCore::AcceleratedEffect::apply):

Canonical link: https://commits.webkit.org/299031@main