Move ENABLE_SWIFTUI flag to PlatformEnableCocoa.h https://bugs.webkit.org/show_bug.cgi?id=297028 rdar://157706818
Reviewed by Richard Robinson.

It was previously only defined in xcconfigs, which is an anti-pattern.
It's still useful to have an xcconfig definition in WebKit, because a
handful of build settings are keyed off of $(WK_ENABLE_SWIFTUI), but the
TestWebKitAPI usage can read directly from WTF.

With the definition moved, change -[WKWebView
_updateScrollGeometryWithContentOffset:contentSize:] to be a no-op in
builds where the SwiftUI API does not exist. This avoids confusion in
SPI checking by calling respondsToSelector for an internal delegate
methods that does not exist.

* Source/WTF/wtf/PlatformEnableCocoa.h:
* Source/WebKit/Configurations/Base.xcconfig:
* Source/WebKit/UIProcess/API/Cocoa/WKWebView.mm:

(-[WKWebView _updateScrollGeometryWithContentOffset:contentSize:]):
* Tools/TestWebKitAPI/Configurations/TestWebKitAPIBundle.xcconfig:
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKScrollGeometryTests.mm: Only
run these tests when ENABLE_SWIFTUI is active.

Canonical link: https://commits.webkit.org/299034@main