[WPE][GTK] API test runner: allow to pass a list of tests and subtests to run. https://bugs.webkit.org/show_bug.cgi?id=297751
Reviewed by Carlos Garcia Campos.

The API-Tests EWS bots have been not working properly for GTK and WPE since 288955@main
which introduced an optimization to run only the subtests that failed with the patch
on the retry step that runs without the patch.

The problem is that the API test runner script for GTK and WPE is not understanding the
syntax. So it is not running anything when the list of tests and subtests to run is passed.

This patch modifies it to support receiving a list of tests and subtests to run (previously
it could receive a list of tests, but it only allowed to specify the list of subtests when
running one test alone). The format used is MainTestName:subtest-name.

This also improves the behaviour around some corner cases, like when there is an exception
running the binary test, then the python runner generates a fake subtest with names like
'afterAll' or 'beforeAll' that the binary runner won't understand if passed to it as a subtest
name. This patch modifies the python runner to detects that and in that case opt to run all
the subtests for that given test.

* Tools/glib/api_test_runner.py:
(TestRunner.__init__):
(TestRunner._get_tests):
(TestRunner._get_test_short_name):
(TestRunner):
(TestRunner._getsubtests_to_run_for_test):
(TestRunner.run_tests):
(TestRunner.run_tests.generate_test_list_for_json_output):
* Tools/glib/glib_test_runner.py:
(GLibTestRunner.run):

Canonical link: https://commits.webkit.org/299049@main