Fix spill / fill order of vector registers in wasmPrologueSIMD. https://bugs.webkit.org/show_bug.cgi?id=297753 rdar://158708738
Reviewed by Yijia Huang.

This is a speculative fix for a performance regression after 298634@main.

Looking at the generated LLIntAssembly.h code that spill vector registers, ...

Before 298634@main, the spill looks like:
    "str q0, [x29, #-112] \n"            // WebAssembly.asm:283
    "str q1, [x29, #-128] \n"
    "str q2, [x29, #-144] \n"
    "str q3, [x29, #-160] \n"
    "str q4, [x29, #-176] \n"
    "str q5, [x29, #-192] \n"
    "str q6, [x29, #-208] \n"
    "str q7, [x29, #-224] \n"

After 298634@main, the spill looks like:
    "str q1, [x29, #-128] \n"            // WebAssembly.asm:328
    "str q0, [x29, #-112] \n"            // WebAssembly.asm:329
    "str q3, [x29, #-160] \n"            // WebAssembly.asm:328
    "str q2, [x29, #-144] \n"            // WebAssembly.asm:329
    "str q5, [x29, #-192] \n"            // WebAssembly.asm:328
    "str q4, [x29, #-176] \n"            // WebAssembly.asm:329
    "str q7, [x29, #-224] \n"            // WebAssembly.asm:328
    "str q6, [x29, #-208] \n"            // WebAssembly.asm:329

We can see that the 298634@main unexpectedly scrambled the spill order so that it is no longer
storing consecutively in memory.  This patch restores the original order.  This is done both
for the spill and the matching fill.

* Source/JavaScriptCore/llint/WebAssembly.asm:

Canonical link: https://commits.webkit.org/299054@main