Fix test 'http/tests/websocket/tests/hybi/upgrade-simple-ws.html' https://bugs.webkit.org/show_bug.cgi?id=158480
Reviewed by Alexey Proskuryakov.

This test had several errors. First one, it was accessing the property 'ws.URL',
when it should be 'ws.url'. The URL attribute of a WebSocket object is
lowercase.

The other error was a very short timeout. Many other tests in the same
folder have a timeout of 3000ms, while this test had a timeout of 300ms.

* LayoutTests/TestExpectations:
* LayoutTests/http/tests/websocket/tests/hybi/upgrade-simple-ws-expected.txt:
* LayoutTests/http/tests/websocket/tests/hybi/upgrade-simple-ws.html:

Canonical link: https://commits.webkit.org/299057@main