CSP: Link header with rel=preload does not recognize nonces https://bugs.webkit.org/show_bug.cgi?id=222484 rdar://75060055
Reviewed by Ryosuke Niwa.

Erroneous CSP violations can happen if a CSP report-only header has a nonce and we attempt a preload
that contains the nonce.

When building a preload request we aren't copying the nonce from the link header or element's
nonce attribute into the fetch options we use when creating the resource request. So later when
the loader tries to validate we're allowed to do the load there's no nonce to check against and we
can trip over a Content-Security-Policy-Report-Only header and generate an erroneous violation report.

* LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/link-preload-from-header-report-only-nonce.sub-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/link-preload-from-header-report-only-nonce.sub.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/link-preload-from-header-report-only-nonce.sub.html.sub.headers: Added.
* LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/link-preload-report-only-nonce.sub-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/link-preload-report-only-nonce.sub.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/link-preload-report-only-nonce.sub.html.sub.headers: Added.
* LayoutTests/imported/w3c/web-platform-tests/content-security-policy/support/pass.js: Added.
* Source/WebCore/html/parser/HTMLResourcePreloader.cpp:
(WebCore::PreloadRequest::resourceRequest):

Canonical link: https://commits.webkit.org/299070@main