ASSERTION FAILED: InjectedBundle::singleton().isTestRunning() is constantly crashing fast/eventsource/eventsource-attribute-listeners.html https://bugs.webkit.org/show_bug.cgi?id=297141 rdar://157879724
Reviewed by Alexey Proskuryakov.

fast/eventsource/eventsource-attribute-listeners.html is crashing in
EWS, not because of any fault of its own, but because it usually runs in
succession to fast/events/touch/page-scaled-touch-gesture-click.html.

The bad test calls an asynchronous function in the body onload handler,
i.e. does not await. This can put the test infrastructure in a bad
state. We receive InjectedBundle::done() early because the test is so
quickly done with and as a result the injected bundle clears out its
held TestRunner instance. However, TestRunner has a positive refcount
because of InjectedBundlePage::frameDidChangeLocation() --> dump(). We
end up in this code path because the test runner is _not_ currently
"waiting for done", which we should have, had the test properly run its
runTest() method. Finally, when the runTest() method does eventually
finish executing, we reach TestRunner::waitUntilDone() (since it still
has a positive refcount), and crash at the release assert because the
injected bundle has already cleared out its test runner instance.

In the good case, we never end up having a high refcount through dump()
because the test runner would have been "waiting for done" already, but
more importantly, we don't think the test is done because the async test
code is actually awaited on. As a result, we don't accidentally wipe the
test runner instance held by the injected bundle singleton.

While the TestRunner being kept alive may seem confusing, the ref counts
are never unbalanced, and it's fair for TestRunner callers to keep it
alive. What's unfortunate is that we go against the standard practice
from elsewhere in TestRunner and just release assert over the absence of
a test runner instance in the injected bundle. Everywhere else just
early returns from the work it is supposed to do. This patch follows
suit so that tests do not unnecessarily crash if they're poorly written.
We do add an error log to surface some information.

To address the test's badness, we first extricate the async function call
into a proper 'load' event handler. The test still times out, though, as
we already know in rdar://80387340. This patch also does the bare minimum
to move the test from a time out to a consistent failure by moving away
from unuspported EventSender API.

Finally, we make two test expectation changes:

1. Remove the crashing expectation for fast/eventsource/eventsource-attribute-listeners.html
2. Remove the 'failing' expectation because fast/events/touch/page-scaled-touch-gesture-click.html
   still relies on touch event support which is missing from open source
   configurations, and hence should be skipped altogether, like all the
   other tests under fast/events/touch.

* LayoutTests/fast/events/touch/page-scaled-touch-gesture-click.html:
* LayoutTests/platform/glib/TestExpectations:
* LayoutTests/platform/ios/TestExpectations:
* Tools/WebKitTestRunner/InjectedBundle/TestRunner.cpp:
(WTR::TestRunner::waitUntilDone):

Canonical link: https://commits.webkit.org/299084@main