Sync `whitespace-integer.html` from Blink / Chromium source https://bugs.webkit.org/show_bug.cgi?id=297840
Reviewed by Nikolas Zimmermann.

We use to only pass 27 tests out of 192 but we progressed at 299083@main.
This patch is to sync test so we can track any future regressions, I was also
trying to merge this test here [1] but I thought to just sync it separately.

[1] https://github.com/WebKit/WebKit/pull/44323

* LayoutTests/svg/parser/whitespace-integer.html:
* LayoutTests/svg/parser/whitespace-integer-expected.txt:

Canonical link: https://commits.webkit.org/299102@main