[align-content] out-of-flow box with no sibling ignores align-content https://bugs.webkit.org/show_bug.cgi?id=297828
Reviewed by Antti Koivisto.

Inline content mixed with out-of-flow boxes generate IFC (inlineLayout()) to
not only size and position inline boxes but also compute the static position
for those out-of-flow boxes.
Now in case of a single out-of-flow box, we still need to compute the static position
but when hasSimpleOutOfFlowContentOnly is true, we don't create the IFC object anymore (fast path).

RenderBlockFlow::shiftForAlignContent aligns the boxes by either calling shiftLinesBy on IFC (and svg) or moving the _block_ child boxes when !childrenInline().
The logic here misses the "single out-of-flow box with fast path" case, where
childrenInline() is true but there's no IFC (or svg) inline object to call.
Let's re-shuffle the conditions in shiftForAlignContent to cover all cases.

* LayoutTests/imported/w3c/web-platform-tests/css/css-align/abspos/align-out-of-flow-only-content-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-align/abspos/align-out-of-flow-only-content-ref.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-align/abspos/align-out-of-flow-only-content.html: Added.
* Source/WebCore/rendering/RenderBlockFlow.cpp:
(WebCore::RenderBlockFlow::shiftForAlignContent):

Canonical link: https://commits.webkit.org/299107@main