[WebDriver][BiDi] Fix error forwarding for errors with detail string https://bugs.webkit.org/show_bug.cgi?id=294824
Reviewed by BJ Burg.

Ensure we check if we're getting an error message with a detail string,
using it for the "message" field in the outgoing BiDi error message.

This avoids getting generic "Unknown Error" with messages
"InvalidParameter;The parameter foo is invalid" if using the
"...AND_DETAILS" macros.

* Source/WebKit/UIProcess/Automation/WebDriverBidiProcessor.cpp:
(WebKit::WebDriverBidiProcessor::sendBidiMessage):

Canonical link: https://commits.webkit.org/299108@main