Add id and lang attributes to WebVTT root node and add layout tests for lang selectors https://bugs.webkit.org/show_bug.cgi?id=297796 rdar://158961777
Reviewed by Jer Noble.

This patch adds a lang attribute to the webvtt root node matching the track's srclang attribute,
if it has one. This is a step towards implementing the behavior that a cue should be matched by
::cue([lang=xy]), where xy is the track srclang.

This patch also adds an id attribute to the webvtt root node if the node is given one in the .vtt.
This is a step towards implementing the ::cue(#id) selector.

Lastly this patch adds missing tests for ::cue([lang=xy]) and ::cue(lang). As of now,
lang_color.html (which tests ::cue(lang)) passes but lang_attribute.html (tests ::cue([lang=xy]))
does not, due to work that still needs to be done to make ::cue(selector) match the root webvtt node.

* LayoutTests/imported/w3c/web-platform-tests/webvtt/rendering/cues-with-video/processing-model/selectors/cue_function/lang_object/lang_attribute-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/webvtt/rendering/cues-with-video/processing-model/selectors/cue_function/lang_object/lang_attribute.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/webvtt/rendering/cues-with-video/processing-model/selectors/cue_function/lang_object/lang_color-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/webvtt/rendering/cues-with-video/processing-model/selectors/cue_function/lang_object/lang_color.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/webvtt/rendering/cues-with-video/processing-model/support/test_lang_object.vtt: Added.
* Source/WebCore/html/track/VTTCue.cpp:
(WebCore::VTTCue::obtainCSSBoxes):

Canonical link: https://commits.webkit.org/299115@main