[iOS] TestWebKitAPI.WritingTools.CompositionWithAttemptedEditing is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=297877 rdar://158255565
Reviewed by Richard Robinson.

* Tools/TestWebKitAPI/Tests/WebKitCocoa/WritingTools.mm:
(TEST(WritingTools, CompositionWithAttemptedEditing)):

In order for text replacement to finish, the text animation must also finish.
The current 0.1s wait is not deterministically sufficient in order for the animation
to complete on newer OSes. Bump it up to 1s to ensure the replacement is done.

Note that the FIXME here is copied from other WritingTools tests which have the
same architectural limitation.

Canonical link: https://commits.webkit.org/299140@main