[GTK][WPE] Use the IPC connection to send/receive credentials https://bugs.webkit.org/show_bug.cgi?id=297846
Reviewed by Michael Catanzaro.

We currently use a dedicated socket pair connection for that, but we
could just use the IPC connection. Also, we only need to do that when
spawning sandboxed web processes, for other process types or when
sandbox is disabled we already have the pid, so we don't need to slow
down the process launch.

Canonical link: https://commits.webkit.org/299151@main