Rename the key-pair elements returned in the gcryptGenerateX25519Keys method https://bugs.webkit.org/show_bug.cgi?id=289693
Reviewed by Carlos Garcia Campos.

The gcryptGenerateX25519Keys method returns a std::pair of KeyMaterial variables
which should contain the public and private key. The caller expects the first
element to be the public key.

We were returning the data in the wrong order, so among other issues, it caused
that the public key generated by the JWK export of the private key doesn't match
the original key-pair's public key material.

This change fixes the issue by switching the variable names used to hold the
private and public key pair. This is also more consistent with the nomenclature
used in the Ed25519 generateKey() method..

Regarding tests, this change make the successes_X25519.https.any.html tests to
pass, so updating also the platform-specific TestExpectations file.

* LayoutTests/platform/glib/TestExpectations:
* Source/WebCore/crypto/gcrypt/CryptoKeyOKPGCrypt.cpp:
(WebCore::gcryptGenerateX25519Keys):

Canonical link: https://commits.webkit.org/299198@main