REGRESSION (298383@main): Clicking inside text selection does not reset the selection if the coordinates of the event location are not whole https://bugs.webkit.org/show_bug.cgi?id=297930 rdar://159141822
Reviewed by Abrar Rahman Protyasha.

When a mouse release event occurs inside a text selection, we clear the selection if
the mouse hasn't moved since the last mouse press event. However, starting in
298383@main, the location of the last mouse press event stored in m_dragStartPosition
was always floored, while the mouse release event location was not. This resulted in
the selection failing to clear unless the x and y coordinates of the release event were
both whole. Fix this by flooring both values during the comparison.

Added support for moving the mouse to non-integral coordinates in WKTR so that this
behavior can be tested in new layout test `selection-clear-on-click-fractional-coordinates.html`.

* LayoutTests/fast/events/selection-clear-on-click-fractional-coordinates-expected.txt: Added.
* LayoutTests/fast/events/selection-clear-on-click-fractional-coordinates.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerlock/pointerevent_coordinates_when_locked-expected.txt:
* LayoutTests/media/modern-media-controls/slider/slider-value.html:
* LayoutTests/media/modern-media-controls/volume-support/volume-support-click.html:
* LayoutTests/media/modern-media-controls/volume-support/volume-support-drag.html:
* LayoutTests/platform/ios/TestExpectations:
* LayoutTests/platform/mac-wk1/TestExpectations:
* Source/WebCore/page/EventHandler.cpp:
(WebCore::EventHandler::handleMouseReleaseEvent):
* Tools/WebKitTestRunner/InjectedBundle/Bindings/EventSendingController.idl:
* Tools/WebKitTestRunner/InjectedBundle/EventSendingController.cpp:
(WTR::EventSendingController::mouseMoveTo):
(WTR::EventSendingController::asyncMouseMoveTo):
* Tools/WebKitTestRunner/InjectedBundle/EventSendingController.h:

Canonical link: https://commits.webkit.org/299237@main