[WPE] WPE Platform: build WPEPlatform as a part of libWPEWebKit https://bugs.webkit.org/show_bug.cgi?id=298014
Reviewed by Adrian Perez de Castro.

Building a separate shared library for WPEPlatform means we end up with
two copies of WTF which has caused some problems with globals. We still
install a different pkg-config file for WPEPlatform but pointing to
libWPEWebKit instead, similar to what we do with the WPEPlatform
libraries.

Canonical link: https://commits.webkit.org/299246@main