[WebXR][OpenXR] Hand interaction extension is never enabled https://bugs.webkit.org/show_bug.cgi?id=298018
Reviewed by Carlos Garcia Campos.

The hand interaction extension is never enabled because it was never
detected as available even if the runtime supported it. The problem
was that the code was not properly passing a span to the method that
checks availability. Instead it was passing a char*.

* Source/WebKit/UIProcess/XR/openxr/PlatformXROpenXR.cpp:
(WebKit::OpenXRCoordinator::createInstance):

Canonical link: https://commits.webkit.org/299249@main