TestWebKitAPI.ScreenTime.DoNotDonateURLsInOccludedWebView is a constant timeout https://bugs.webkit.org/show_bug.cgi?id=297982 rdar://157154868
Reviewed by Aditya Keerthi.

Set the `NSEventConcurrentProcessingEnabled` default to NO as a workaround to
fix the timeout of the `DoNotDonateURLsInOccludedWebView` API test. While this
is considered a workaround due to the current issue, the change is actually the
original behavior, which will be explained below.

Concurrent event processing was always enabled by AppKit. AppKit used to start
concurrent event processing under the condition that the application finished launching.
However, TestWebKitAPI is not an actual NSApplication. Thus, the
`NSApplicationDidFinishLaunchingNotification` would never be fired, and
concurrent event processing is not started.

Recent AppKit + HIToolbox changes now start concurrent event processing regardless
if an application finished launching or not, and consequently starts an NSEventThread.

Concurrent event processing currently breaks window occlusion detection in apps not using
`NSApplicationMain`, causing the API test to timeout when waiting for the window to be occluded.
By setting the default to NO, concurrent event processing is effectively disabled,
making the behavior the same prior to the AppKit + HIToolbox changes.

* Tools/TestWebKitAPI/mac/mainMac.mm:
(handleArguments):

Canonical link: https://commits.webkit.org/299278@main